+++ CHANGELOG.md
@@ -15,6 +15,7 @@ For a full diff see [`2.1.0...main`][2.1.0...main].
 - Renamed `Content` to `BodyMatter` and `Parsed::content()` to `Parsed::bodyMatter()` ([#398]), by [@localheinz]
 - Extracted `Content` as a value object ([#399]), by [@localheinz]
 - Renamed `Parsed` to `ParsedContent` ([#400]), by [@localheinz]
+- Renamed `Content` to `UnparsedContent` ([#407]), by [@localheinz]
 
 ## [`2.1.0`][2.1.0]
 
@@ -108,5 +109,6 @@ For a full diff see [`4e97e14...0.1.0`][4e97e14...0.1.0].
 [#398]: https://github.com/ergebnis/front-matter/pull/398
 [#399]: https://github.com/ergebnis/front-matter/pull/399
 [#400]: https://github.com/ergebnis/front-matter/pull/400
+[#407]: https://github.com/ergebnis/front-matter/pull/407
 
 [@localheinz]: https://github.com/localheinz